feat: 데이터 저장, 삭제 디바운싱 추가

+++ lib/presentation/my_info/my_info_model.dart
@@ -1,11 +1,16 @@
+import 'dart:async';
+
 import 'package:firebase_auth/firebase_auth.dart' as firebase_auth;
 import 'package:flutter/material.dart';
 import 'package:flutter_flash_card/data/remote/back_up_service.dart';
 
 class MyInfoModel extends ChangeNotifier {
+  Timer? _debounceTimer;
+
   final BackUpService _backUpService;
   final firebase_auth.FirebaseAuth _firebaseAuth;
   List<String> _backUpList = [];
+  bool _isLoading = false;
 
   List<String> get backUpList => _backUpList;
 
@@ -28,23 +33,51 @@ class MyInfoModel extends ChangeNotifier {
   }
 
   Future<void> uploadToCloud() async {
-    final currentUser = _firebaseAuth.currentUser;
+    if (_isLoading) return;
 
-    if (currentUser != null) {
-      await _backUpService.backupToFirestore(currentUser.uid);
-      await loadBackUpList();
-    }
-    debugPrint('---upload---');
+    _debounceTimer?.cancel();
+    _debounceTimer = Timer(const Duration(milliseconds: 300), () async {
+      _isLoading = true;
+      notifyListeners();
+
+      final currentUser = _firebaseAuth.currentUser;
+
+      if (currentUser != null) {
+        await _backUpService.backupToFirestore(currentUser.uid);
+        await loadBackUpList();
+      }
+      debugPrint('---upload---');
+
+      _isLoading = false;
+      notifyListeners();
+    });
   }
 
   Future<void> deleteBackUpData() async {
-    final currentUser = _firebaseAuth.currentUser;
+    if (_isLoading) return;
 
-    if (currentUser != null) {
-      await _backUpService.deleteBackups(currentUser.uid);
-      _backUpList = [];
-    }
+    _debounceTimer?.cancel();
+    _debounceTimer = Timer(const Duration(milliseconds: 300), () async {
+      _isLoading = true;
+      notifyListeners();
 
-    notifyListeners();
+      final currentUser = _firebaseAuth.currentUser;
+
+      if (currentUser != null) {
+        await _backUpService.deleteBackups(currentUser.uid);
+        _backUpList = [];
+      }
+
+      notifyListeners();
+
+      _isLoading = false;
+      notifyListeners();
+    });
+  }
+
+  @override
+  void dispose() {
+    _debounceTimer?.cancel();
+    super.dispose();
   }
 }
